+++ modules/persistence/lib/schemas/feature-toggle-schema.js
@@ -6,6 +6,8 @@ var Schema = mongoose.Schema;
 var FeatureToggle = new Schema({
   name: String,
   description: String,
+  // Hidden from the UI but still accessible via API
+  hidden: Boolean,
   criteria: Schema.Types.Mixed
 });
 FeatureToggle.schemaTypeName = 'FeatureToggle';