+++ public/js/vue/left-menu/components/list-item.vue
@@ -1,9 +1,13 @@
 <script>
 import { mapActions } from 'vuex';
+import LoadingSpinner from '../../components/loading-spinner.vue';
 import parseForTemplate from 'gitter-web-shared/parse/left-menu-primary-item';
 
 export default {
   name: 'ListItem',
+  components: {
+    LoadingSpinner
+  },
   props: {
     item: {
       type: Object,
@@ -16,6 +20,10 @@ export default {
         // The value must match one of these strings
         return ['all', 'org'].indexOf(value) !== -1;
       }
+    },
+    active: {
+      type: Boolean,
+      default: false
     }
   },
   computed: {
@@ -37,10 +45,15 @@ export default {
 </script>
 
 <template>
-  <li class="list-item" :class="{ 'is-favourited': item.favourite }">
+  <li
+    class="list-item"
+    :class="{ 'is-favourited': item.favourite }"
+    :data-favourite="item.favourite"
+  >
     <a
       ref="link"
       class="list-item-link"
+      :class="{ 'is-active': active }"
       :href="item.absoluteRoomUri"
       @click.stop.prevent="roomClick(item)"
     >
@@ -55,6 +68,8 @@ export default {
         <span v-else-if="nameData.displayName" class="name-piece">{{ nameData.displayName }}</span>
         <span v-else class="name-piece">{{ nameData.name }}</span>
       </h2>
+
+      <loading-spinner v-if="item.loading" />
 
       <div v-if="item.mentions" class="mention-indicator">@</div>
       <div v-else-if="item.unreadItems" class="unread-indicator">
@@ -104,6 +119,7 @@ export default {
 
   text-decoration: none;
 
+  &.is-active,
   &:hover,
   &:focus {
     cursor: pointer;