+++ server/handlers/renderers/vue/mixin-vue-left-menu-data.js
@@ -26,11 +26,19 @@ async function mixinHbsDataForVueLeftMenu(req, existingData) {
   const uriContext = req.uriContext;
   const room = uriContext && uriContext.troupe;
 
+  const isMobile = req.isPhone;
+
   const vueLeftMenuHtmlOutput = await vueRenderToString({
     moduleToRender: 'left-menu',
     storeData: {
+      isMobile,
+      isLoggedIn: !!user,
+
       roomMap,
-      displayedRoomId: room && room.id
+      displayedRoomId: room && room.id,
+
+      leftMenuPinnedState: !isMobile,
+      leftMenuExpandedState: false
     }
   });
 
@@ -38,6 +46,7 @@ async function mixinHbsDataForVueLeftMenu(req, existingData) {
     ...existingData,
 
     layout: 'chat-layout',
+    isMobile,
     useVueLeftMenu: useVueLeftMenu,
     leftMenuHtml: vueLeftMenuHtmlOutput,