+++ test/public-js/vue/left-menu/components/list-item-test.js
@@ -67,6 +67,35 @@ describe('list-item', () => {
     expect(wrapper.element).toMatchSnapshot();
   });
 
+  it('favourite room matches snapshot', () => {
+    factory({
+      item: {
+        ...createSerializedRoomFixture('my-community/community'),
+        favourite: 1
+      }
+    });
+    expect(wrapper.element).toMatchSnapshot();
+  });
+
+  it('active room matches snapshot', () => {
+    factory({
+      item: createSerializedRoomFixture('my-community/room1'),
+      active: true
+    });
+    expect(wrapper.element).toMatchSnapshot();
+  });
+
+  it('favourite loading room matches snapshot', () => {
+    factory({
+      item: {
+        ...createSerializedRoomFixture('my-community/community'),
+        favourite: 1,
+        loading: true
+      }
+    });
+    expect(wrapper.element).toMatchSnapshot();
+  });
+
   it('calls store action "changeDisplayedRoom" and "toggleLeftMenu" when item is clicked', () => {
     const room = createSerializedRoomFixture('my-community/community');
     factory({